refactor: introduce report utils package and use it in cli apply
