--- cmd/cli/kubectl-kyverno/apply/.gitignore
@@ -1,1 +1,0 @@
-/disallow_latest_tag.yaml
